func: add reload and clean up after removing a dependency
