+++ nomad/structs/structs.go
@@ -10959,160 +10959,3 @@ func NewRpcError(err error, code *int64) *RpcError {
 func (r *RpcError) Error() string {
 	return r.Message
 }
-
-type JobDependency struct {
-	Name   string
-	Status string
-}
-
-func (d *JobDependency) Equal(o *JobDependency) bool {
-	if d == nil || o == nil {
-		return d == o
-	}
-
-	return d == o ||
-		d.Name == o.Name &&
-			d.Status == o.Status
-}
-
-func (d *JobDependency) Validate() error {
-	if d == nil {
-		return errors.New("dependency job block is required")
-	}
-
-	if d.Name == "" {
-		return errors.New("dependency job name is mandatory")
-	}
-
-	if d.Status == "" {
-		return errors.New("dependency job status is mandatory")
-	}
-
-	return nil
-}
-
-func (d *JobDependency) Canonicalize() {
-	if d == nil {
-		return
-	}
-
-	if d.Status == "" {
-		d.Status = "completed"
-	}
-}
-
-func (d *JobDependency) String() string {
-	if d == nil {
-		return ""
-	}
-
-	return fmt.Sprintf("%s: %s", d.Name, d.Status)
-}
-
-// A Dependency is used to restrict placement options.
-type Dependency struct {
-	Timeout         time.Duration
-	ActionOnTimeout string
-	Jobs            []*JobDependency
-}
-
-// Equal checks if two dependencies are equal.
-func (d *Dependency) Equal(o *Dependency) bool {
-	if d == nil || o == nil {
-		return d == o
-	}
-
-	if len(d.Jobs) != len(o.Jobs) {
-		return false
-	}
-
-	jEqual := true
-	for i := range d.Jobs {
-		jEqual = jEqual && d.Jobs[i].Equal(o.Jobs[i])
-	}
-
-	return d == o ||
-		d.Timeout == o.Timeout &&
-			d.ActionOnTimeout == o.ActionOnTimeout &&
-			jEqual
-}
-
-func (d *Dependency) Copy() *Dependency {
-	if d == nil {
-		return nil
-	}
-
-	jobs := make([]*JobDependency, 0, len(d.Jobs))
-	for _, job := range d.Jobs {
-		if job == nil {
-			jobs = append(jobs, nil)
-			continue
-		}
-
-		copy := *job
-		jobs = append(jobs, &copy)
-	}
-
-	return &Dependency{
-		Timeout:         d.Timeout,
-		ActionOnTimeout: d.ActionOnTimeout,
-		Jobs:            jobs,
-	}
-}
-
-func (d *Dependency) String() string {
-	jobs := make([]string, 0, len(d.Jobs))
-	for _, j := range d.Jobs {
-		jobs = append(jobs, j.String())
-	}
-
-	return fmt.Sprintf("%s %s: %s", d.Timeout, d.ActionOnTimeout, strings.Join(jobs, ", "))
-}
-
-func (d *Dependency) Validate() error {
-	var mErr multierror.Error
-	if d == nil {
-		return nil
-	}
-
-	if d.Timeout <= 0 {
-		mErr.Errors = append(mErr.Errors, errors.New("Missing or invalid timeout in dependency"))
-	}
-
-	if d.ActionOnTimeout == "" {
-		mErr.Errors = append(mErr.Errors, errors.New("Missing action_on_timeout in dependency"))
-	} else if d.ActionOnTimeout != "reject" {
-		mErr.Errors = append(mErr.Errors, fmt.Errorf("Invalid action_on_timeout in dependency: %q", d.ActionOnTimeout))
-	}
-
-	if len(d.Jobs) == 0 {
-		mErr.Errors = append(mErr.Errors, errors.New("Missing job in dependency"))
-	}
-
-	for idx, job := range d.Jobs {
-		if err := job.Validate(); err != nil {
-			mErr.Errors = append(mErr.Errors, fmt.Errorf("Dependency job %d validation failed: %v", idx+1, err))
-		}
-	}
-
-	return mErr.ErrorOrNil()
-}
-
-func (d *Dependency) Canonicalize() {
-	if d == nil {
-		return
-	}
-
-	if d.ActionOnTimeout == "" {
-		d.ActionOnTimeout = "reject"
-	}
-
-	for _, job := range d.Jobs {
-		job.Canonicalize()
-	}
-}
-
-// DiffID fulfills the DiffableWithID interface.
-func (d *Dependency) DiffID() string {
-	return d.String()
-}
