+++ nomad/worker.go
@@ -623,7 +623,7 @@ func (w *Worker) invokeScheduler(snap *state.StateSnapshot, eval *structs.Evalua
 	// Create the scheduler, or use the special core scheduler
 	var sched sstructs.Scheduler
 	if eval.Type == structs.JobTypeCore {
-		sched = NewCoreScheduler(w.srv, snap, w, scheduler.WithDependencyChecker(w.srv.dependencyCoordinator))
+		sched = NewCoreScheduler(w.srv, snap, w, withCoreDependencyChecker(w.srv.dependencyCoordinator))
 	} else {
 		sched, err = scheduler.NewScheduler(eval.Type, w.logger, w.srv.workersEventCh,
 			snap, w, scheduler.WithDependencyChecker(w.srv.dependencyCoordinator))
